pyproject.toml
[project]
name = "seobi-db"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "flask>=3.1.1",
    "flask-cors>=6.0.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
]
